+++ src/logging/configuration.ts
@@ -10,26 +10,48 @@ const serviceContext = {
 
 /**
  * Options that should be passed when initializing a `pino` logger.
- * It ensures the message will be contained in the `message` key. The `serviceContext` key will also be added.
+ *
+ * - It ensures the message will be contained in the `message` key.
+ * - The `serviceContext` key is added.
+ * - The `req.headers.authorization` key is redacted in all log objects.
  */
 let pinoConfiguration: LoggerOptions = {
   messageKey: 'message',
   base: { serviceContext },
+  redact: { paths: ['req.headers.authorization'] },
 };
 
 /**
  * Updates the pino configuration returned by {@link getPinoConfiguration}.
- * New options will be merged with the existing ones. The {@link LoggerOptions.base} bindings will be merged.
+ * New options will be merged with the existing ones.
+ * The {@link LoggerOptions.base} bindings will be merged, as well as the {@link LoggerOptions.redact.paths}.
  *
  * @param options Custom pino options to include in the existing configuration.
  */
 export function updatePinoConfiguration(options: Partial<LoggerOptions>): void {
+  const newRedactOptions = Array.isArray(options.redact)
+    ? { paths: options.redact }
+    : options.redact;
+  const mergedRedactPaths = [
+    ...new Set([
+      ...(Array.isArray(pinoConfiguration.redact)
+        ? pinoConfiguration.redact
+        : pinoConfiguration.redact?.paths ?? []),
+      ...(newRedactOptions?.paths ?? []),
+    ]),
+  ];
+
   pinoConfiguration = {
     ...pinoConfiguration,
     ...options,
     base: {
       ...pinoConfiguration.base,
       ...options.base,
+    },
+    redact: {
+      ...pinoConfiguration.redact,
+      ...newRedactOptions,
+      paths: mergedRedactPaths,
     },
   };
 }